+++ titiler/pgstac/mosaic.py
@@ -26,10 +26,11 @@ from rio_tiler.mosaic import mosaic_reader
 from rio_tiler.tasks import create_tasks, filter_tasks
 from rio_tiler.types import AssetInfo, BBox
 
-from titiler.pgstac.settings import CacheSettings, RetrySettings
+from titiler.pgstac.settings import CacheSettings, PgstacSettings, RetrySettings
 from titiler.pgstac.utils import retry
 
 cache_config = CacheSettings()
+pgstac_config = PgstacSettings()
 retry_config = RetrySettings()
 
 
@@ -273,11 +274,13 @@ class PGSTACBackend(BaseBackend):
             "include": ["assets", "id", "bbox", "collection"],
         }
 
-        scan_limit = scan_limit or 10000
-        items_limit = items_limit or 100
-        time_limit = time_limit or 5
-        exitwhenfull = True if exitwhenfull is None else exitwhenfull
-        skipcovered = True if skipcovered is None else skipcovered
+        scan_limit = scan_limit or pgstac_config.scan_limit
+        items_limit = items_limit or pgstac_config.items_limit
+        time_limit = time_limit or pgstac_config.time_limit
+        exitwhenfull = (
+            pgstac_config.exitwhenfull if exitwhenfull is None else exitwhenfull
+        )
+        skipcovered = pgstac_config.skipcovered if skipcovered is None else skipcovered
 
         with self.pool.connection() as conn:
             with conn.cursor() as cursor: